// apps/web/src/components/puck/pageui/front-centre/product-feature-alt/types.ts
export interface PageUiProductFeatureAltProps {
  title: string;
  description: string;
  showCta: boolean;
  ctaLabel: string;
  ctaHref: string;
  showCtaNote: boolean;
  ctaNote: string;
  showImage: boolean;
  imageSrc: string;
  variant: 'primary' | 'secondary';
  withBackground: boolean;
  withBackgroundGlow: boolean;
  sectionId: string;
}